Move create page to landing page
