+++ frontend/src/components/config/App.tsx
@@ -6,7 +6,6 @@ import NotFound from '../../views/NotFound';
 import { CustomRoute } from './Route';
 import GamePage from '../../views/Game';
 import GameLayout from '../layout/Game';
-import CreateGamePage from '../../views/Create';
 import JoinGamePage from '../../views/Join';
 import GameResultsPage from '../../views/Results';
 
@@ -15,7 +14,6 @@ function App() {
     <Switch>
       <CustomRoute path="/" component={LandingPage} layout={MainLayout} exact />
       <CustomRoute path="/game" component={GamePage} layout={GameLayout} exact />
-      <CustomRoute path="/game/create" component={CreateGamePage} layout={MainLayout} exact />
       <CustomRoute path="/game/join" component={JoinGamePage} layout={MainLayout} exact />
       <CustomRoute path="/game/results" component={GameResultsPage} layout={GameLayout} exact />
       <CustomRoute path="*" component={NotFound} layout={MainLayout} />
